more test correcting

+++ src/test-c/test-ok/sup_db_t.c
@@ -1,7 +1,7 @@
-double main(){
+int main(){
     double i,j,k;
     j = 3.3;
     k = 2.1;
     i = (j>k);
-    return i-1;
+    return (int) i-1;
 }
